+++ news-bundle/src/Resources/contao/modules/ModuleNews.php
@@ -10,6 +10,8 @@
 
 namespace Contao;
 
+use FOS\HttpCache\ResponseTagger;
+
 /**
  * Parent class for news modules.
  *
@@ -77,7 +79,6 @@ abstract class ModuleNews extends Module
 	 */
 	protected function parseArticle($objArticle, $blnAddArchive=false, $strClass='', $intCount=0)
 	{
-		/** @var FrontendTemplate|object $objTemplate */
 		$objTemplate = new \FrontendTemplate($this->news_template);
 		$objTemplate->setData($objArticle->row());
 
@@ -214,6 +215,15 @@ abstract class ModuleNews extends Module
 				$this->import($callback[0]);
 				$this->{$callback[0]}->{$callback[1]}($objTemplate, $objArticle->row(), $this);
 			}
+		}
+
+		// Tag the response
+		if (System::getContainer()->has('fos_http_cache.http.symfony_response_tagger'))
+		{
+			/** @var ResponseTagger $responseTagger */
+			$responseTagger = System::getContainer()->get('fos_http_cache.http.symfony_response_tagger');
+			$responseTagger->addTags(array('contao.db.tl_news.' . $objArticle->id));
+			$responseTagger->addTags(array('contao.db.tl_news_archive.' . $objArticle->pid));
 		}
 
 		return $objTemplate->parse();