+++ news-bundle/src/Resources/contao/modules/ModuleNewsArchive.php
@@ -42,9 +42,7 @@ class ModuleNewsArchive extends ModuleNews
 	{
 		if (TL_MODE == 'BE')
 		{
-			/** @var BackendTemplate|object $objTemplate */
 			$objTemplate = new \BackendTemplate('be_wildcard');
-
 			$objTemplate->wildcard = '### ' . Utf8::strtoupper($GLOBALS['TL_LANG']['FMD']['newsarchive'][0]) . ' ###';
 			$objTemplate->title = $this->headline;
 			$objTemplate->id = $this->id;
@@ -166,7 +164,7 @@ class ModuleNewsArchive extends ModuleNews
 
 				// Get the current page
 				$id = 'page_a' . $this->id;
-				$page = (\Input::get($id) !== null) ? \Input::get($id) : 1;
+				$page = \Input::get($id) ?? 1;
 
 				// Do not index or cache the page if the page number is outside the range
 				if ($page < 1 || $page > max(ceil($total/$this->perPage), 1))