+++ news-bundle/tests/EventListener/InsertTagsListenerTest.php
@@ -41,9 +41,11 @@ class InsertTagsListenerTest extends ContaoTestCase
         ];
 
         $framework = $this->mockContaoFramework($adapters);
-        $listener = new InsertTagsListener($framework);
 
-        $this->assertSame('http://localhost/share/news.xml', $listener->onReplaceInsertTags('news_feed::2'));
+        $listener = new InsertTagsListener($framework);
+        $url = $listener->onReplaceInsertTags('news_feed::2', false, null, []);
+
+        $this->assertSame('http://localhost/share/news.xml', $url);
     }
 
     public function testReplacesTheNewsTags(): void
@@ -78,17 +80,17 @@ class InsertTagsListenerTest extends ContaoTestCase
 
         $this->assertSame(
             '<a href="news/foo-is-not-bar.html" title="&quot;Foo&quot; is not &quot;bar&quot;">"Foo" is not "bar"</a>',
-            $listener->onReplaceInsertTags('news::2')
+            $listener->onReplaceInsertTags('news::2', false, null, [])
         );
 
         $this->assertSame(
             '<a href="news/foo-is-not-bar.html" title="&quot;Foo&quot; is not &quot;bar&quot;">',
-            $listener->onReplaceInsertTags('news_open::2')
+            $listener->onReplaceInsertTags('news_open::2', false, null, [])
         );
 
         $this->assertSame(
             'news/foo-is-not-bar.html',
-            $listener->onReplaceInsertTags('news_url::2')
+            $listener->onReplaceInsertTags('news_url::2', false, null, [])
         );
 
         $this->assertSame(
@@ -98,12 +100,12 @@ class InsertTagsListenerTest extends ContaoTestCase
 
         $this->assertSame(
             '&quot;Foo&quot; is not &quot;bar&quot;',
-            $listener->onReplaceInsertTags('news_title::2')
+            $listener->onReplaceInsertTags('news_title::2', false, null, [])
         );
 
         $this->assertSame(
             '<p>Foo does not equal bar.</p>',
-            $listener->onReplaceInsertTags('news_teaser::2')
+            $listener->onReplaceInsertTags('news_teaser::2', false, null, [])
         );
     }
 
@@ -111,7 +113,7 @@ class InsertTagsListenerTest extends ContaoTestCase
     {
         $listener = new InsertTagsListener($this->mockContaoFramework());
 
-        $this->assertFalse($listener->onReplaceInsertTags('link_url::2'));
+        $this->assertFalse($listener->onReplaceInsertTags('link_url::2', false, null, []));
     }
 
     public function testReturnsAnEmptyStringIfThereIsNoModel(): void
@@ -123,7 +125,7 @@ class InsertTagsListenerTest extends ContaoTestCase
 
         $listener = new InsertTagsListener($this->mockContaoFramework($adapters));
 
-        $this->assertSame('', $listener->onReplaceInsertTags('news_feed::3'));
-        $this->assertSame('', $listener->onReplaceInsertTags('news_url::3'));
+        $this->assertSame('', $listener->onReplaceInsertTags('news_feed::3', false, null, []));
+        $this->assertSame('', $listener->onReplaceInsertTags('news_url::3', false, null, []));
     }
 }