+++ news-bundle/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Contao news bundle change log
+
+## DEV
+
+ * Show the 404 page if an external news is opened in the news reader (see #33).