+++ news-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Contao news bundle change log
 
-## DEV
+## 4.6.0-RC1 (2018-07-16)
 
  * Link to the news article if no image link has been defined (see #30).
  * Show the 404 page if an external news is opened in the news reader (see #33).